requirements.txt
pyOpenSSL==19.1.0
Jinja2==2.11.2
huepy==1.2.1
pycryptodome==3.9.7
PyYAML==5.3.1
secrets==1.0.2
